+++ CHANGELOG.md
@@ -7,6 +7,8 @@ zuora_client CHANGELOG
 - Remove test files that pass. This helps clean up search
 - Add pagination support to CatalogAPI
 - Fix serializer for get_product_type
+- Add ADDITIONAL_INFO_OPTIONAL__c, PlatformID__c, and CustomerType__c to POSTAccountType
+- Add deleter for run_billing, invoice, and collect in POSTAccountType
 
 2.0.0
 -----